Adds support for input[type=file] for enctype=multipart/form-data
My hypothesis here is that when the user needs to use an
input[type=file] they will also need to use a form which is has the
enctype `multipart/form-data` (and also makes the assumption that
the form method is not 'GET', because a form upload should not be mapped
to a GET resource). This modifies the `serializeForm` method to add two
different accumulators: one for collecting the values of the input
fields as string values and then encoding them as a stringified
application/x-www-form-urlencoded URL, and one which collects them as a
`FormData` object which can then be directly set as the body of the http
request.

The new serializeForm method is still backwards compatible because it
uses a StringParams() accumulator by default.

Note: I haven't tested this implementation very carefully (yet), so it
should be carefully reviewed and tested before merging.